refactor(ansi): rewrite ansi and writer
