+++ src/color/ansi_writer_hyperlink_test.go
@@ -19,7 +19,7 @@ func TestGenerateHyperlinkNoUrl(t *testing.T) {
 		{Text: "sample text with no url", ShellName: shell.BASH, Expected: "sample text with no url"},
 	}
 	for _, tc := range cases {
-		a := Ansi{}
+		a := AnsiWriter{}
 		a.Init(tc.ShellName)
 		hyperlinkText := a.GenerateHyperlink(tc.Text)
 		assert.Equal(t, tc.Expected, hyperlinkText)
@@ -52,7 +52,7 @@ func TestGenerateHyperlinkWithUrl(t *testing.T) {
 		},
 	}
 	for _, tc := range cases {
-		a := Ansi{}
+		a := AnsiWriter{}
 		a.Init(tc.ShellName)
 		hyperlinkText := a.GenerateHyperlink(tc.Text)
 		assert.Equal(t, tc.Expected, hyperlinkText)
@@ -70,35 +70,10 @@ func TestGenerateHyperlinkWithUrlNoName(t *testing.T) {
 		{Text: "[](http://www.google.be)", ShellName: shell.BASH, Expected: "[](http://www.google.be)"},
 	}
 	for _, tc := range cases {
-		a := Ansi{}
+		a := AnsiWriter{}
 		a.Init(tc.ShellName)
 		hyperlinkText := a.GenerateHyperlink(tc.Text)
 		assert.Equal(t, tc.Expected, hyperlinkText)
-	}
-}
-
-func TestFormatText(t *testing.T) {
-	cases := []struct {
-		Case     string
-		Text     string
-		Expected string
-	}{
-		{Case: "single format", Text: "This <b>is</b> white", Expected: "This \x1b[1mis\x1b[22m white"},
-		{Case: "double format", Text: "This <b>is</b> white, this <b>is</b> orange", Expected: "This \x1b[1mis\x1b[22m white, this \x1b[1mis\x1b[22m orange"},
-		{Case: "underline", Text: "This <u>is</u> white", Expected: "This \x1b[4mis\x1b[24m white"},
-		{Case: "italic", Text: "This <i>is</i> white", Expected: "This \x1b[3mis\x1b[23m white"},
-		{Case: "strikethrough", Text: "This <s>is</s> white", Expected: "This \x1b[9mis\x1b[29m white"},
-		{Case: "dimmed", Text: "This <d>is</d> white", Expected: "This \x1b[2mis\x1b[22m white"},
-		{Case: "flash", Text: "This <f>is</f> white", Expected: "This \x1b[5mis\x1b[25m white"},
-		{Case: "reversed", Text: "This <r>is</r> white", Expected: "This \x1b[7mis\x1b[27m white"},
-		{Case: "double", Text: "This <i><f>is</f></i> white", Expected: "This \x1b[3m\x1b[5mis\x1b[25m\x1b[23m white"},
-		{Case: "overline", Text: "This <o>is</o> white", Expected: "This \x1b[53mis\x1b[55m white"},
-	}
-	for _, tc := range cases {
-		a := Ansi{}
-		a.InitPlain()
-		formattedText := a.formatText(tc.Text)
-		assert.Equal(t, tc.Expected, formattedText, tc.Case)
 	}
 }
 
@@ -114,7 +89,7 @@ func TestGenerateFileLink(t *testing.T) {
 		{Text: `[Windows](file:C:/Windows)`, Expected: "\x1b]8;;file:C:/Windows\x1b\\Windows\x1b]8;;\x1b\\"},
 	}
 	for _, tc := range cases {
-		a := Ansi{}
+		a := AnsiWriter{}
 		a.Init(shell.PWSH)
 		hyperlinkText := a.GenerateHyperlink(tc.Text)
 		assert.Equal(t, tc.Expected, hyperlinkText)
