+++ src/engine/engine_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/jandedobbeleer/oh-my-posh/color"
-	"github.com/jandedobbeleer/oh-my-posh/console"
 	"github.com/jandedobbeleer/oh-my-posh/mock"
 	"github.com/jandedobbeleer/oh-my-posh/platform"
 	"github.com/jandedobbeleer/oh-my-posh/shell"
@@ -71,15 +70,16 @@ func TestPrintPWD(t *testing.T) {
 			Env:   make(map[string]string),
 			Shell: "shell",
 		})
-		ansi := &color.Ansi{}
-		ansi.InitPlain()
+
+		writer := &color.AnsiWriter{}
+		writer.Init(shell.GENERIC)
 		engine := &Engine{
 			Env: env,
 			Config: &Config{
 				PWD:   tc.PWD,
 				OSC99: tc.OSC99,
 			},
-			Ansi: ansi,
+			Writer: writer,
 		}
 		engine.printPWD()
 		got := engine.print()
@@ -101,25 +101,16 @@ func engineRender() {
 	cfg := LoadConfig(env)
 	defer testClearDefaultConfig()
 
-	ansi := &color.Ansi{}
-	ansi.InitPlain()
 	writerColors := cfg.MakeColors()
 	writer := &color.AnsiWriter{
-		Ansi:               ansi,
 		TerminalBackground: shell.ConsoleBackgroundColor(env, cfg.TerminalBackground),
 		AnsiColors:         writerColors,
 	}
-	consoleTitle := &console.Title{
-		Env:      env,
-		Ansi:     ansi,
-		Template: cfg.ConsoleTitleTemplate,
-	}
+	writer.Init(shell.GENERIC)
 	engine := &Engine{
-		Config:       cfg,
-		Env:          env,
-		Writer:       writer,
-		ConsoleTitle: consoleTitle,
-		Ansi:         ansi,
+		Config: cfg,
+		Env:    env,
+		Writer: writer,
 	}
 
 	engine.PrintPrimary()
@@ -128,5 +119,129 @@ func engineRender() {
 func BenchmarkEngineRenderPalette(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		engineRender()
+	}
+}
+
+func TestGetTitle(t *testing.T) {
+	cases := []struct {
+		Template      string
+		Root          bool
+		User          string
+		Cwd           string
+		PathSeparator string
+		ShellName     string
+		Expected      string
+	}{
+		{
+			Template:      "{{.Env.USERDOMAIN}} :: {{.PWD}}{{if .Root}} :: Admin{{end}} :: {{.Shell}}",
+			Cwd:           "C:\\vagrant",
+			PathSeparator: "\\",
+			ShellName:     "PowerShell",
+			Root:          true,
+			Expected:      "\x1b]0;MyCompany :: C:\\vagrant :: Admin :: PowerShell\a",
+		},
+		{
+			Template:      "{{.Folder}}{{if .Root}} :: Admin{{end}} :: {{.Shell}}",
+			Cwd:           "C:\\vagrant",
+			PathSeparator: "\\",
+			ShellName:     "PowerShell",
+			Expected:      "\x1b]0;vagrant :: PowerShell\a",
+		},
+		{
+			Template:      "{{.UserName}}@{{.HostName}}{{if .Root}} :: Admin{{end}} :: {{.Shell}}",
+			Root:          true,
+			User:          "MyUser",
+			PathSeparator: "\\",
+			ShellName:     "PowerShell",
+			Expected:      "\x1b]0;MyUser@MyHost :: Admin :: PowerShell\a",
+		},
+	}
+
+	for _, tc := range cases {
+		env := new(mock.MockedEnvironment)
+		env.On("Pwd").Return(tc.Cwd)
+		env.On("Home").Return("/usr/home")
+		env.On("PathSeparator").Return(tc.PathSeparator)
+		env.On("TemplateCache").Return(&platform.TemplateCache{
+			Env: map[string]string{
+				"USERDOMAIN": "MyCompany",
+			},
+			Shell:    tc.ShellName,
+			UserName: "MyUser",
+			Root:     tc.Root,
+			HostName: "MyHost",
+			PWD:      tc.Cwd,
+			Folder:   "vagrant",
+		})
+		ansi := &color.AnsiWriter{}
+		ansi.Init(shell.GENERIC)
+		engine := &Engine{
+			Config: &Config{
+				ConsoleTitleTemplate: tc.Template,
+			},
+			Writer: ansi,
+			Env:    env,
+		}
+		title := engine.getTitleTemplateText()
+		got := ansi.FormatTitle(title)
+		assert.Equal(t, tc.Expected, got)
+	}
+}
+
+func TestGetConsoleTitleIfGethostnameReturnsError(t *testing.T) {
+	cases := []struct {
+		Template      string
+		Root          bool
+		User          string
+		Cwd           string
+		PathSeparator string
+		ShellName     string
+		Expected      string
+	}{
+		{
+			Template:      "Not using Host only {{.UserName}} and {{.Shell}}",
+			User:          "MyUser",
+			PathSeparator: "\\",
+			ShellName:     "PowerShell",
+			Expected:      "\x1b]0;Not using Host only MyUser and PowerShell\a",
+		},
+		{
+			Template:      "{{.UserName}}@{{.HostName}} :: {{.Shell}}",
+			User:          "MyUser",
+			PathSeparator: "\\",
+			ShellName:     "PowerShell",
+			Expected:      "\x1b]0;MyUser@ :: PowerShell\a",
+		},
+		{
+			Template: "\x1b[93m[\x1b[39m\x1b[96mconsole-title\x1b[39m\x1b[96m ≡\x1b[39m\x1b[31m +0\x1b[39m\x1b[31m ~1\x1b[39m\x1b[31m -0\x1b[39m\x1b[31m !\x1b[39m\x1b[93m]\x1b[39m",
+			Expected: "\x1b]0;[console-title ≡ +0 ~1 -0 !]\a",
+		},
+	}
+
+	for _, tc := range cases {
+		env := new(mock.MockedEnvironment)
+		env.On("Pwd").Return(tc.Cwd)
+		env.On("Home").Return("/usr/home")
+		env.On("TemplateCache").Return(&platform.TemplateCache{
+			Env: map[string]string{
+				"USERDOMAIN": "MyCompany",
+			},
+			Shell:    tc.ShellName,
+			UserName: "MyUser",
+			Root:     tc.Root,
+			HostName: "",
+		})
+		ansi := &color.AnsiWriter{}
+		ansi.Init(shell.GENERIC)
+		engine := &Engine{
+			Config: &Config{
+				ConsoleTitleTemplate: tc.Template,
+			},
+			Writer: ansi,
+			Env:    env,
+		}
+		title := engine.getTitleTemplateText()
+		got := ansi.FormatTitle(title)
+		assert.Equal(t, tc.Expected, got)
 	}
 }
