+++ src/shell/constants.go
@@ -1,12 +1,12 @@
 package shell
 
 const (
-	ZSH   = "zsh"
-	BASH  = "bash"
-	PWSH  = "pwsh"
-	FISH  = "fish"
-	PWSH5 = "powershell"
-	CMD   = "cmd"
-	NU    = "nu"
-	PLAIN = "shell"
+	ZSH     = "zsh"
+	BASH    = "bash"
+	PWSH    = "pwsh"
+	FISH    = "fish"
+	PWSH5   = "powershell"
+	CMD     = "cmd"
+	NU      = "nu"
+	GENERIC = "shell"
 )
